requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pydantic-settings>=2.4.0
python-dotenv
httpx
falkordb>=1.0.0
# Install graphiti-core from fork with FalkorDB support
git+https://github.com/vlad29042/graphiti.git@master